+++ infrastructure/driven-adapters/r2dbc-mysql/src/main/java/co/com/pragma/solicitud/r2dbcmysql/entity/R2dbcTipoPrestamo.java
@@ -16,6 +16,6 @@ public class R2dbcTipoPrestamo {
     private String nombre;
     private Long montoMinimo;
     private Long montoMaximo;
-    private int tasaInteres;
+    private Double tasaInteres;
     private boolean validacionAutomatica;
 }
